requirements.txt
annotated-types==0.6.0
apparmor==3.1.6
application-utility==1.3.2
asgiref==3.7.2
attrs==23.1.0
autocommand==2.2.2
btrfsutil==6.6.3
catfish==4.18.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==41.0.7
CT3==3.3.3
dbus-python==1.3.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.2
django-allauth==0.54.0
docopt==0.6.2
fastjsonschema==2.19.1
filelock==3.12.0
gufw==24.4.0
idna==3.5
inflect==7.0.0
jaraco.context==4.3.0
jaraco.functools==3.9.0
jaraco.text==3.11.1
keyutils==0.6
lensfun==0.3.4
LibAppArmor==3.1.6
libtorrent==2.0.9
lightdm-gtk-greeter-settings==1.2.2
lit==16.0.6.dev0
manjaro-sdk==0.1.1
menulibre==2.2.3
more-itertools==10.1.0
mugshot==0.4.3
netsnmp-python==1.0a1
nftables==0.1
npyscreen==4.10.5
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.2
pacman_mirrors==4.24.0
pexpect==4.9.0
platformdirs==4.1.0
ply==3.11
psutil==5.9.6
ptyprocess==0.7.0
pycairo==1.25.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyGObject==3.46.0
PyJWT==2.7.0
pymongo==4.6.1
PySide6==6.6.1
python3-openid==3.2.0
pyxdg==0.28
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
shiboken2==5.15.11
shiboken6==6.6.1
shiboken6-generator==6.6.1
sqlparse==0.4.4
systemd-python==235
tomli==2.0.1
trove-classifiers==2023.12.1
typing_extensions==4.9.0
udiskie==2.5.1
ufw==0.36.2
urllib3==1.26.18
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
virtualenv==20.23.0
